requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
Authlib==1.4.0
black==23.10.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
decorator==5.1.1
Django==4.2.7
django-cors-headers==4.3.0
django-crontab==0.7.1
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
executing==2.0.1
idna==3.10
inflection==0.5.1
ipython==8.17.2
jedi==0.19.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
pathspec==0.11.2
platformdirs==3.11.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
pycparser==2.22
Pygments==2.16.1
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.32.3
rpds-py==0.12.0
setuptools==75.6.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.13.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.9
whitenoise==6.6.0
